Небольщое изменение дизайна карты, добавлен деб пакет и загрузчик
